+++ src/components/loadMoreBtn.js
@@ -1,0 +1,2 @@
+export const loadMoreBtnLayout = () =>
+  `<button class="load-more" type="button">load more</button>`;
