--- src/layouts/loadMoreBtn.js
@@ -1,4 +1,0 @@
-const loadMoreBtnLayout = () =>
-  `<button class="load-more" type="button">load more</button>`;
-
-export default loadMoreBtnLayout;
